Improves question loading by saving questions in a session variable everytime a new lesson is opened.
The session arrays ['current_lesson'] and ['questions'] are introduced to keep track of which lesson the user is doing and whether it is necessary to query from the database again. As long as the user is doing questions from the same lesson, the database is only queried once. If the user chooses to do questions from another lesson, then the database will be queried for the new questions.